+++ src/datarefinery/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2026 Pointmatic
 # SPDX-License-Identifier: Apache-2.0
 
-__version__ = "0.19.0"
+__version__ = "0.21.0"
 
 from datarefinery.core.datarefinery import DataRefinery, materialize, resolve_instance
 from datarefinery.core.instance import Instance